mmc: sdhci: Use quirks2 for defining new quirks
Revert back quirks to 32 bits long and use quirks2 to define any new
required quirks.

Bug 1330567

Reviewed-on: http://git-master/r/260564
(cherry picked from commit 836c16ce754c154856369cc2d5494ece77aa573c)

Reviewed-on: http://git-master/r/274873
Tested-by: Mrutyunjay Sawant <msawant@nvidia.com>

+++ drivers/mmc/host/sdhci-pltfm.h
@@ -1,5 +1,6 @@
 /*
  * Copyright 2010 MontaVista Software, LLC.
+ * Copyright (c) 2013, NVIDIA CORPORATION.  All rights reserved.
  *
  * Author: Anton Vorontsov <avorontsov@ru.mvista.com>
  *
@@ -17,7 +18,7 @@
 
 struct sdhci_pltfm_data {
 	struct sdhci_ops *ops;
-	u64 quirks;
+	u32 quirks;
 	u32 quirks2;
 };
